command execution to its own module
