+++ src/extension.ts
@@ -9,7 +9,7 @@ const markdownItAttrs = require('markdown-it-attrs');
 import {readKatapodConfig} from './configuration';
 import {getWorkingDir} from './filesystem';
 import {log} from './logging';
-// import {sendText, sendTextsPerTerminal, runInitScripts} from './runCommands';
+import {sendTextToTerminal, sendTextsPerTerminal, runInitScripts} from './runCommands';
 
 
 let kpEnvironment: any = {
@@ -24,80 +24,18 @@ let kpEnvironment: any = {
 	}
 };
 
-export function sendTextToTerminal (cbContent: any, env: any) {
-	// pick target terminal, with care and fallbacks
-	const targetTerminalTag: string = (cbContent.runSettings || {}).terminal;
-	let targetTerminal: vscode.Terminal;
-	if (targetTerminalTag){
-		targetTerminal = env.components.terminalMap[targetTerminalTag];
-		if (!targetTerminal){
-			targetTerminal = env.components.terminals[0];
-			log('debug', `sendText fails by name and falls to first`);
-		}else{
-			log('debug', `sendText picks terminal by name ${targetTerminalTag}`);
-		}
-	}else{
-		targetTerminal = env.components.terminals[0];
-		log('debug', `sendText picks first term by default`);
-	}
-
-	// run the command
-	targetTerminal.sendText(cbContent.command);
-	vscode.commands.executeCommand('notifications.clearAll');
-}
-
-export function sendTextsPerTerminal(commandMap: any, env: any, logContext: string){
-	Object.entries(commandMap).forEach(([terminalTag, command]) => {
-		log('info', `${logContext}: calling "${command}" on terminal "${terminalTag}"`);
-		sendTextToTerminal(
-            {
-                runSettings: {
-                    terminal: terminalTag,
-                },
-                command,
-            },
-			env
-        );
-	});
-}
-
-export function runInitScripts (env: any) {
-	let config = env.configuration;
-	if (!config.startup){
-		// legacy mode:
-		const waitsh = vscode.Uri.file(path.join(getWorkingDir(), 'wait.sh'));
-		vscode.workspace.fs.stat(waitsh).then(
-			function(){
-				log('debug', 'Executing wait.sh... (legacy mode)');
-				sendTextToTerminal(
-                    {
-                        runSettings: {
-                            terminal: 'cqlsh',
-                        },
-                        command: './wait.sh',
-                    },
-					env
-                );
-			}, 
-			function () {log('debug', 'Skipping wait, wait.sh not found (legacy mode).');}
-		);
-	} else {
-		const startupScripts = config.startup || {};
-		// this maps terminal tags to startup commands to execute there
-		sendTextsPerTerminal(startupScripts, env, "startup");
-	}
-}
-
-function sendText(cbContent: any) {
+// closing over the kpEnvironment to supply a one-arg command to registerCommand
+function sendTextClosure(cbContent: any) {
 	sendTextToTerminal(cbContent, kpEnvironment);
 }
 
 export async function activate(context: vscode.ExtensionContext) {
 	/*
-	Nothing seems to prevent this function from being async (which helps fighting nondeterminism in these calls).
+	Nothing seems to prevent this function from being async
+	(which helps fighting nondeterminism in these calls while avoiding chains of ".then(...)" with all await's).
 	See https://stackoverflow.com/questions/64640967/can-a-vscode-extension-activate-method-be-async.
 	*/
-	context.subscriptions.push(vscode.commands.registerCommand('katapod.sendText', sendText));
+	context.subscriptions.push(vscode.commands.registerCommand('katapod.sendText', sendTextClosure));
 	context.subscriptions.push(vscode.commands.registerCommand('katapod.loadPage', loadPage));
 	context.subscriptions.push(vscode.commands.registerCommand('katapod.reloadPage', reloadPage));
 	context.subscriptions.push(vscode.commands.registerCommand('katapod.start', start));
